+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+
+### Added
+
+- Example: display joint frames from the URDF description
+
 ## [1.2.0] - 2023/03/14
 
 ### Added